flask/src/requirements.txt
alembic==1.13.1
antiorm==1.2.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
db==0.1.1
db-sqlite3==0.0.1
flask==3.0.0
flask-login==0.6.3
flask-migrate==4.0.5
flask-sqlalchemy==3.1.1
idna==3.6
itsdangerous==2.1.2
jinja2==3.1.2
mako==1.3.0
markupsafe==2.1.3
numpy==1.26.2
packaging==23.2
pandas==2.1.4
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlalchemy==2.0.24
tzdata==2023.4
urllib3==2.1.0
werkzeug==3.0.1
gunicorn==20.1.0
openpyxl==3.0.9

xlrd==2.0.1